fix: ensure global listener event types are translated everywhere
The Global User Task Listener event types were previously translated in
the add and edit modals, but not in the list view.
This refactoring ensures common logic is used everywhere (including the
list view) to convert the list of event types to a user-friendly string:
- if `all` is in the list, a specific "All events" is shown, regardless
of other event types in the list
- if `all` is not in the list, the other event types are translated and
joined by commas (,)
