+++ identity/client/src/pages/global-task-listeners/modals/EditModal.tsx
@@ -18,7 +18,11 @@ import {
   LISTENER_EVENT_TYPES,
 } from "src/utility/api/global-task-listeners";
 import { useNotifications } from "src/components/notifications";
-import { LISTENER_TYPE_PATTERN } from "src/pages/global-task-listeners";
+import {
+  getEventTypeLabel,
+  getEventTypeLabels,
+  LISTENER_TYPE_PATTERN,
+} from "src/pages/global-task-listeners";
 import type {
   CreateGlobalTaskListenerRequestBody,
   GlobalTaskListener,
@@ -108,20 +112,6 @@ const EditModal: FC<UseEntityModalProps<GlobalTaskListener>> = ({
     setValue("eventTypes", selectedItems);
   };
 
-  const getEventTypeLabel = (
-    eventType: GlobalTaskListener["eventTypes"][number],
-  ): string => {
-    const labels: Record<GlobalTaskListener["eventTypes"][number], string> = {
-      all: t("eventTypeAll"),
-      creating: t("eventTypeCreating"),
-      updating: t("eventTypeUpdating"),
-      assigning: t("eventTypeAssigning"),
-      completing: t("eventTypeCompleting"),
-      canceling: t("eventTypeCanceling"),
-    };
-    return labels[eventType];
-  };
-
   const onSubmit = async (data: CreateGlobalTaskListenerRequestBody) => {
     const eventTypes = data.eventTypes.includes("all")
       ? ["all" as const]
@@ -206,9 +196,7 @@ const EditModal: FC<UseEntityModalProps<GlobalTaskListener>> = ({
             titleText={t("eventType")}
             label={
               field.value.length > 0
-                ? field.value.includes("all")
-                  ? t("eventTypeAll")
-                  : field.value.map(getEventTypeLabel).join(", ")
+                ? getEventTypeLabels(field.value, t)
                 : t("selectEventTypes")
             }
             items={[...LISTENER_EVENT_TYPES]}
@@ -221,7 +209,7 @@ const EditModal: FC<UseEntityModalProps<GlobalTaskListener>> = ({
               handleEventTypeChange(selectedItems);
             }}
             itemToString={(item: GlobalTaskListenerEventType) =>
-              getEventTypeLabel(item)
+              getEventTypeLabel(item, t)
             }
             invalid={!!fieldState.error}
             invalidText={fieldState.error?.message}
